Address review feedback: dedupe board setup, add semihost exit
- Extract the two boards' shared setup (syscall driver lookup,
  KernelResources, capsule/process init) into a new qemu_arm_mps2_lib
  crate, generic over cortexm::CortexMVariant. Each board's main.rs
  shrinks to just the concrete CortexM3/CortexM4 type argument, plus a
  static_init!() and a #[panic_handler] that structurally can't be
  generic (a static can't reference an enclosing generic function's own
  type parameter).
- Add cortexm::support::semihost_command(), since no ARM semihosting
  primitive existed in-tree; call it from each board's panic handler and
  pass -semihosting to QEMU, so a panic now makes qemu-system-arm exit on
  its own instead of hanging until killed.
- Trim both READMEs down to reference documentation: move LED
  observability into the peripheral list itself, drop verification
  transcripts and the watchdog test narrative (that belongs in commit
  history, not a permanent README), and fix an385/an386 chip_layout.ld
  comments (an386's had copy-pasted "AN385").
- Minor doc-comment cleanups in spi.rs/watchdog.rs.
