+++ arch/cortex-m/src/support.rs
@@ -184,3 +184,50 @@ pub fn is_interrupt_context() -> bool {
 pub fn is_interrupt_context() -> bool {
     unimplemented!()
 }
+
+/// Issue an ARM semihosting call.
+///
+/// `operation` is the semihosting operation number (e.g. `0x18` for
+/// `SYS_EXIT`) and `parameter` is its operation-specific argument. Only
+/// meaningful when running under a semihosting host (e.g. QEMU started with
+/// `-semihosting`, or an attached debug probe); otherwise the `bkpt`
+/// instruction traps with no host to service it.
+#[cfg(any(doc, all(target_arch = "arm", target_os = "none")))]
+#[inline(always)]
+pub unsafe fn semihost_command(operation: u32, parameter: u32) -> u32 {
+    use core::arch::asm;
+    let result;
+
+    // # Safety
+    //
+    // - INPUTS: r0/r1 are set to `operation`/`parameter`, per the ABI ARM
+    //   semihosting defines (ARM's "Semihosting for AArch32 and AArch64"
+    //   specification).
+    // - OUTPUTS: r0 is overwritten with the semihosting call's result.
+    // - Options set:
+    //   - nostack: This does not use the stack.
+    // - Options not set:
+    //   - nomem: not guaranteed in general -- some operations (e.g.
+    //     `SYS_WRITEC`) dereference `parameter` as a pointer.
+    //   - pure, readonly: not applicable, as above.
+    //   - preserves_flags: not documented by the semihosting spec.
+    //   - noreturn: we do fall through (there may be no host to service
+    //     this call at all, e.g. real hardware with no debugger attached).
+    //   - att_syntax: not on arm.
+    //   - raw: not required.
+    unsafe {
+        asm!(
+            "bkpt #0xAB",
+            inout("r0") operation => result,
+            in("r1") parameter,
+            options(nostack),
+        );
+    }
+    result
+}
+
+/// Mock implementation for tests on Travis-CI.
+#[cfg(not(any(doc, all(target_arch = "arm", target_os = "none"))))]
+pub unsafe fn semihost_command(_operation: u32, _parameter: u32) -> u32 {
+    unimplemented!()
+}
