+++ boards/qemu_arm_mps2_an386/chip_layout.ld
@@ -2,14 +2,17 @@
 /* SPDX-License-Identifier: Apache-2.0 OR MIT                         */
 /* Copyright Tock Contributors 2026.                                  */
 
-/* Memory layout for the QEMU ARM MPS2 AN385 (Cortex-M3) machine.
+/* Memory layout for the QEMU ARM MPS2 AN386 (Cortex-M4) machine.
  *
  * QEMU maps code/flash (SSRAM1) at 0x0, up to a hard 4 MiB cap
  * (`armv7m_load_kernel(..., 0, 0x400000)` in `hw/arm/mps2.c`), and RAM at
  * 0x21000000, backed by a fixed 16 MiB (`default_ram_size` in the same
- * file; QEMU errors if `-m` is overridden). We only claim a modest slice
- * of each; QEMU backs whatever the linker script declares regardless of
- * the machine's full capacity.
+ * file; QEMU errors if `-m` is overridden).
+ *
+ * We only expose a more limited portion of the SRAM, to better reflect
+ * the constraints of typical chips. However, QEMU's emulation does not
+ * enforce the limited range here (i.e., it will always emulate the full 16 MiB
+ * of SRAM and allow accesses outside the range we specify here).
  *
  * rom = 256KB (kernel)
  * prog = 256KB (apps)
