+++ boards/qemu_arm_mps2_lib/Cargo.toml
@@ -1,0 +1,22 @@
+# Licensed under the Apache License, Version 2.0 or the MIT License.
+# SPDX-License-Identifier: Apache-2.0 OR MIT
+# Copyright Tock Contributors 2026.
+
+[package]
+name = "qemu_arm_mps2_lib"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+
+[dependencies]
+components = { path = "../components" }
+cortexm = { path = "../../arch/cortex-m" }
+kernel = { path = "../../kernel" }
+qemu_arm_mps2_chip = { path = "../../chips/qemu_arm_mps2_chip" }
+
+capsules-core = { path = "../../capsules/core" }
+capsules-extra = { path = "../../capsules/extra" }
+capsules-system = { path = "../../capsules/system" }
+
+[lints]
+workspace = true
