+++ boards/qemu_arm_mps2_lib/src/lib.rs
@@ -1,0 +1,300 @@
+// Licensed under the Apache License, Version 2.0 or the MIT License.
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+// Copyright Tock Contributors 2026.
+
+//! Shared kernel setup for the QEMU ARM MPS2 AN385 (Cortex-M3) and AN386
+//! (Cortex-M4) boards.
+//!
+//! The two boards are identical other than their CPU core, so this crate is
+//! generic over the [`CortexMVariant`] and provides everything that
+//! genericity allows to be shared: the `Platform` syscall driver lookup, the
+//! `ChipHw` type, and the process/capsule setup in [`start()`]. What can't
+//! be made generic stays in each board's own `main.rs`/`io.rs`:
+//! - The `#[panic_handler]` function and its `PANIC_RESOURCES` static, since
+//!   a `static` can't be generic and `#[panic_handler]` requires a concrete,
+//!   non-generic function.
+//! - `kernel::stack_size!`, an item-position macro that can't be invoked
+//!   from inside a generic function body.
+//! - The board name used in the boot banner.
+
+#![no_std]
+
+use cortexm::CortexMVariant;
+use kernel::capabilities;
+use kernel::component::Component;
+use kernel::debug::PanicResources;
+use kernel::platform::chip::Chip;
+use kernel::platform::{KernelResources, SyscallDriverLookup};
+use kernel::utilities::single_thread_value::SingleThreadValue;
+use kernel::{create_capability, static_init};
+
+pub const NUM_PROCS: usize = 4;
+
+pub type ChipHw<C> = qemu_arm_mps2_chip::chip::QemuArmMps2Chip<
+    'static,
+    C,
+    qemu_arm_mps2_chip::Mps2DefaultPeripherals<'static>,
+>;
+pub type ProcessPrinterInUse = capsules_system::process_printer::ProcessPrinterText;
+type SchedulerInUse = components::sched::round_robin::RoundRobinComponentType;
+
+pub struct Platform {
+    console: &'static capsules_core::console::Console<'static>,
+    scheduler: &'static SchedulerInUse,
+    systick: cortexm::systick::SysTick,
+    led: &'static capsules_core::led::LedDriver<
+        'static,
+        qemu_arm_mps2_chip::led::Led<'static>,
+        { qemu_arm_mps2_chip::led::NUM_LEDS as usize },
+    >,
+    alarm: &'static capsules_core::alarm::AlarmDriver<
+        'static,
+        capsules_core::virtualizers::virtual_alarm::VirtualMuxAlarm<
+            'static,
+            qemu_arm_mps2_chip::timer::Timer<'static>,
+        >,
+    >,
+    spi: &'static capsules_core::spi_controller::Spi<
+        'static,
+        capsules_core::virtualizers::virtual_spi::VirtualSpiMasterDevice<
+            'static,
+            qemu_arm_mps2_chip::spi::Spi<'static>,
+        >,
+    >,
+    watchdog: &'static qemu_arm_mps2_chip::watchdog::Watchdog,
+}
+
+impl SyscallDriverLookup for Platform {
+    fn with_driver<F, R>(&self, driver_num: usize, f: F) -> R
+    where
+        F: FnOnce(Option<&dyn kernel::syscall::SyscallDriver>) -> R,
+    {
+        match driver_num {
+            capsules_core::console::DRIVER_NUM => f(Some(self.console)),
+            capsules_core::led::DRIVER_NUM => f(Some(self.led)),
+            capsules_core::alarm::DRIVER_NUM => f(Some(self.alarm)),
+            capsules_core::spi_controller::DRIVER_NUM => f(Some(self.spi)),
+            _ => f(None),
+        }
+    }
+}
+
+impl<C: CortexMVariant> KernelResources<ChipHw<C>> for Platform {
+    type SyscallDriverLookup = Self;
+    type SyscallFilter = ();
+    type ProcessFault = ();
+    type Scheduler = SchedulerInUse;
+    type SchedulerTimer = cortexm::systick::SysTick;
+    type WatchDog = qemu_arm_mps2_chip::watchdog::Watchdog;
+    type ContextSwitchCallback = ();
+
+    fn syscall_driver_lookup(&self) -> &Self::SyscallDriverLookup {
+        self
+    }
+    fn syscall_filter(&self) -> &Self::SyscallFilter {
+        &()
+    }
+    fn process_fault(&self) -> &Self::ProcessFault {
+        &()
+    }
+    fn scheduler(&self) -> &Self::Scheduler {
+        self.scheduler
+    }
+    fn scheduler_timer(&self) -> &Self::SchedulerTimer {
+        &self.systick
+    }
+    fn watchdog(&self) -> &Self::WatchDog {
+        self.watchdog
+    }
+    fn context_switch_callback(&self) -> &Self::ContextSwitchCallback {
+        &()
+    }
+}
+
+/// Peripherals and kernel state ready for [`chip()`] and [`finish_start()`].
+pub struct EarlyInit {
+    pub peripherals: &'static qemu_arm_mps2_chip::Mps2DefaultPeripherals<'static>,
+    pub processes: &'static kernel::process::ProcessArray<NUM_PROCS>,
+    pub board_kernel: &'static kernel::Kernel,
+}
+
+/// Runs the CPU-variant init and allocates peripherals/kernel state.
+///
+/// Split out from [`finish_start()`] because the `ChipHw<C>` allocation in
+/// between the two has to be a `static_init!()` written at a concrete,
+/// non-generic call site (a `static` can't reference a generic function's
+/// own type parameter — see each board's `main.rs`), so each board calls:
+/// `early_init`, then its own `static_init!(ChipHw<C>, ...)`, then
+/// `finish_start`.
+///
+/// `panic_resources` is board-owned for the same reason but filled in here
+/// so both boards only have to declare it, not populate it.
+#[inline(never)]
+pub unsafe fn early_init<C: CortexMVariant>(
+    panic_resources: &'static SingleThreadValue<PanicResources<ChipHw<C>, ProcessPrinterInUse>>,
+) -> EarlyInit {
+    ChipHw::<C>::init();
+
+    kernel::deferred_call::initialize_deferred_call_state::<<ChipHw<C> as Chip>::ThreadIdProvider>(
+    );
+
+    let _ = panic_resources
+        .bind_to_thread::<<ChipHw<C> as Chip>::ThreadIdProvider>(PanicResources::new());
+
+    let peripherals = static_init!(
+        qemu_arm_mps2_chip::Mps2DefaultPeripherals<'static>,
+        qemu_arm_mps2_chip::Mps2DefaultPeripherals::new()
+    );
+
+    let processes = components::process_array::ProcessArrayComponent::new()
+        .finalize(components::process_array_component_static!(NUM_PROCS));
+    let board_kernel = static_init!(kernel::Kernel, kernel::Kernel::new(processes.as_slice()));
+
+    EarlyInit {
+        peripherals,
+        processes,
+        board_kernel,
+    }
+}
+
+/// Finishes board setup and starts loading processes.
+///
+/// `chip` must have been allocated by the caller via
+/// `static_init!(ChipHw<C>, ChipHw::<C>::new(early_init.peripherals))`
+/// after [`early_init()`] — see its docs for why.
+#[inline(never)]
+pub unsafe fn finish_start<C: CortexMVariant>(
+    early: EarlyInit,
+    chip: &'static ChipHw<C>,
+) -> (
+    &'static kernel::Kernel,
+    &'static Platform,
+    &'static ChipHw<C>,
+) {
+    let EarlyInit {
+        peripherals,
+        processes,
+        board_kernel,
+    } = early;
+
+    let uart_mux = components::console::UartMuxComponent::new(&peripherals.uart0, 115200)
+        .finalize(components::uart_mux_component_static!());
+
+    let console = components::console::ConsoleComponent::new(
+        board_kernel,
+        capsules_core::console::DRIVER_NUM,
+        uart_mux,
+        create_capability!(capabilities::MemoryAllocationCapability),
+    )
+    .finalize(components::console_component_static!());
+
+    components::debug_writer::DebugWriterComponent::new::<<ChipHw<C> as Chip>::ThreadIdProvider>(
+        uart_mux,
+        create_capability!(capabilities::SetDebugWriterCapability),
+    )
+    .finalize(components::debug_writer_component_static!());
+
+    let alarm_mux = components::alarm::AlarmMuxComponent::new(&peripherals.timer0).finalize(
+        components::alarm_mux_component_static!(qemu_arm_mps2_chip::timer::Timer),
+    );
+
+    let alarm = components::alarm::AlarmDriverComponent::new(
+        board_kernel,
+        capsules_core::alarm::DRIVER_NUM,
+        alarm_mux,
+        create_capability!(capabilities::MemoryAllocationCapability),
+    )
+    .finalize(components::alarm_component_static!(
+        qemu_arm_mps2_chip::timer::Timer
+    ));
+
+    kernel::create_typed_capability!(process_console_cap, ProcessConsoleCap:
+        kernel::capabilities::ProcessManagementCapability,
+        kernel::capabilities::ProcessStartCapability
+    );
+    let process_console = components::process_console::ProcessConsoleComponent::new(
+        board_kernel,
+        uart_mux,
+        alarm_mux,
+        components::process_printer::ProcessPrinterTextComponent::new()
+            .finalize(components::process_printer_text_component_static!()),
+        None,
+        process_console_cap,
+    )
+    .finalize(components::process_console_component_static!(
+        qemu_arm_mps2_chip::timer::Timer,
+        ProcessConsoleCap
+    ));
+    let _ = process_console.start();
+
+    let spi_mux = components::spi::SpiMuxComponent::new(&peripherals.spi_shield0).finalize(
+        components::spi_mux_component_static!(qemu_arm_mps2_chip::spi::Spi),
+    );
+
+    let spi = components::spi::SpiSyscallComponent::new(
+        board_kernel,
+        spi_mux,
+        qemu_arm_mps2_chip::spi::ChipSelect,
+        capsules_core::spi_controller::DRIVER_NUM,
+        create_capability!(capabilities::MemoryAllocationCapability),
+    )
+    .finalize(components::spi_syscall_component_static!(
+        qemu_arm_mps2_chip::spi::Spi
+    ));
+
+    let led = components::led::LedsComponent::new().finalize(components::led_component_static!(
+        qemu_arm_mps2_chip::led::Led<'static>,
+        peripherals.fpgaio.led(0),
+        peripherals.fpgaio.led(1),
+    ));
+
+    let scheduler = components::sched::round_robin::RoundRobinComponent::new(processes)
+        .finalize(components::round_robin_component_static!(NUM_PROCS));
+
+    let platform = static_init!(
+        Platform,
+        Platform {
+            console,
+            scheduler,
+            systick: cortexm::systick::SysTick::new(),
+            led,
+            alarm,
+            spi,
+            watchdog: &peripherals.watchdog,
+        }
+    );
+
+    extern "C" {
+        /// Beginning of the ROM region containing app images.
+        static _sapps: u8;
+        /// End of the ROM region containing app images.
+        static _eapps: u8;
+        /// Beginning of the RAM region for app memory.
+        static mut _sappmem: u8;
+        /// End of the RAM region for app memory.
+        static _eappmem: u8;
+    }
+
+    let process_management_capability =
+        create_capability!(capabilities::ProcessManagementCapability);
+    kernel::process::load_processes(
+        board_kernel,
+        chip,
+        core::slice::from_raw_parts(
+            core::ptr::addr_of!(_sapps),
+            core::ptr::addr_of!(_eapps) as usize - core::ptr::addr_of!(_sapps) as usize,
+        ),
+        core::slice::from_raw_parts_mut(
+            core::ptr::addr_of_mut!(_sappmem),
+            core::ptr::addr_of!(_eappmem) as usize - core::ptr::addr_of!(_sappmem) as usize,
+        ),
+        &capsules_system::process_policies::PanicFaultPolicy {},
+        &process_management_capability,
+    )
+    .unwrap_or_else(|err| {
+        kernel::debug!("Error loading processes!");
+        kernel::debug!("{:?}", err);
+    });
+
+    (board_kernel, platform, chip)
+}
